requirements.txt
Package                       Version
----------------------------- --------
aiohttp                       3.8.4
aiohttp-retry                 2.8.3
aiosignal                     1.3.1
asgiref                       3.6.0
async-timeout                 4.0.2
attrs                         23.1.0
certifi                       2023.5.7
charset-normalizer            3.1.0
coreapi                       2.3.3
coreschema                    0.0.4
Django                        4.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
frozenlist                    1.3.3
idna                          3.4
itypes                        1.2.0
Jinja2                        3.1.2
MarkupSafe                    2.1.3
multidict                     6.0.4
phonenumbers                  8.13.11
Pillow                        9.5.0
pip                           23.1.2
PyJWT                         2.7.0
python-decouple               3.8
pytz                          2023.3
PyYAML                        5.1.2
requests                      2.30.0
setuptools                    49.2.1
sqlparse                      0.4.4
twilio                        8.2.0
tzdata                        2023.3
uritemplate                   4.1.1
urllib3                       2.0.2
yarl                          1.9.2
